frontend/requirements.txt
requests
streamlit
pandas
Pillow
nltk
tensorflow-cpu